Budget for movies on HomeScreen, TODOs added

+++ app/src/main/java/com/tmdb/movieapp/domain/SimplifiedMovie.kt
@@ -3,7 +3,7 @@ package com.tmdb.movieapp.domain
 data class SimplifiedMovie(
     val id: Int,
     val name: String,
-    val budget: Long,
+    val budget: Int,
     val overview: String,
     val poster_path: String?
 )
